+++ packages/expo-modules-core/android/src/main/cpp/types/JNIToJSIConverter.cpp
@@ -54,8 +54,9 @@ jsi::Value convert(
 
 #define CAST_AND_RETURN(type, classId) \
   if (env->IsInstanceOf(unpackedValue, cache->getJClass(#classId).clazz)) { \
-    return convertToJS(rt, jni::static_ref_cast<type>(value)); \
+    return convertToJS(env, rt, jni::static_ref_cast<type>(value)); \
   }
+#define COMMA ,
 
   CAST_AND_RETURN(jni::JDouble, java/lang/Double)
   CAST_AND_RETURN(jni::JInteger, java/lang/Integer)
@@ -69,9 +70,8 @@ jsi::Value convert(
   CAST_AND_RETURN(JSharedObject::javaobject, expo/modules/kotlin/sharedobjects/SharedObject)
   CAST_AND_RETURN(JavaScriptTypedArray::javaobject, expo/modules/kotlin/jni/JavaScriptTypedArray)
 
-  if (env->IsInstanceOf(unpackedValue, cache->getJClass("java/util/List").clazz)) {
-    return convertToJS(rt, jni::static_ref_cast<jni::JList<jobject>>(value));
-  }
+  CAST_AND_RETURN(jni::JMap<jstring COMMA jobject>, java/util/Map)
+  CAST_AND_RETURN(jni::JCollection<jobject>, java/util/Collection)
 
   // Primitives arrays
   CAST_AND_RETURN(jni::JArrayDouble, [D)
@@ -80,6 +80,7 @@ jsi::Value convert(
   CAST_AND_RETURN(jni::JArrayLong, [J)
   CAST_AND_RETURN(jni::JArrayFloat, [F)
 
+#undef COMMA
 #undef CAST_AND_RETURN
 
   return jsi::Value::undefined();